Bug 498907 - IllegalStateException in ProjectionMapping.toImageLine
Do not allow IllegalStateException to propagate outside of
the MatchingCharacterPainter.paintControl method. This change
does not fix the root cause of bug 498907 but makes the symptoms
less severe.

This solution is intended for the 4.6 maintenance stream.
The 4.7 development stream contains a more general protection
mechanism guarding against exceptions in SWT event listeners.

Change-Id: Ic077d88997dbefb76abf9c8aadf6b57100d44518