Add LinkedModeUI.doExit using a DocumentEvent #33
Copy-Paste doesn't trigger Linked Editing IExitPolicy.doExit(...) so, a new default 'doExit()' nethod is added to
the IExitPolicy interface in order to control exiting the Linked Editing Mode on Paste operations.
The method takes a `DocumentEvent` as an argument and is to be used from `DocumentListener` to check
if we should exit the linked mode.

Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>