brackets highlighting should not contains code mining text #47
`MatchingCharacterPainter` is using the `StyledText.getTextBounds()`
method to determine the width of the highlight box to draw. This method
calculates width based on runs of text, and thus includes the method
parameter hints in that width.

This change instead moves to using `GC.textExtent()` to determine the
width of text to highlight, ensuring that only the desired character is
highlighted.