Fix some issues using inline annotations in different viewers
This reduce the coupling between an annotation and the viewer. As a
result, it fixes some issue when an annotation was displayed on several
viewers and the state of the annotation wasn't allowing multiple
consuming viewers.

Most of the internals to allow an annotation to be shown in multiple
viewers are fixed, however, the API still doesn't make it very clear and
is yet to improve: some constructors and method should be removed to not
allow annotation/viewer coupling.

Fixes https://github.com/eclipse-platform/eclipse.platform.text/issues/206